requirements.txt
# PDF 처리
PyPDF2==3.0.1
pdfplumber==0.10.3

# 벡터 임베딩
sentence-transformers==2.2.2
qdrant-client==1.7.0

# LLM API
openai==1.3.0

# 유틸리티
python-dotenv==1.0.0
tqdm==4.66.1
